+++ flashlib.vision.core/src/main/java/com/flash3388/flashlib/vision/analysis/JsonTarget.java
@@ -1,4 +1,4 @@
-package com.flash3388.flashlib.vision.processing.analysis;
+package com.flash3388.flashlib.vision.analysis;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
