+++ flashlib.vision.core/src/main/java/com/flash3388/flashlib/vision/control/SingleThreadVisionControl.java
@@ -12,7 +12,7 @@ import com.flash3388.flashlib.vision.VisionResult;
 import com.flash3388.flashlib.vision.control.event.NewResultEvent;
 import com.flash3388.flashlib.vision.control.event.VisionListener;
 import com.flash3388.flashlib.vision.processing.Processor;
-import com.flash3388.flashlib.vision.processing.analysis.Analysis;
+import com.flash3388.flashlib.vision.analysis.Analysis;
 import com.notifier.Controllers;
 import com.notifier.EventController;
