+++ flashlib.vision.core/src/main/java/com/flash3388/flashlib/vision/processing/VisionPipeline.java
@@ -2,8 +2,8 @@ package com.flash3388.flashlib.vision.processing;
 
 import com.flash3388.flashlib.vision.Pipeline;
 import com.flash3388.flashlib.vision.VisionException;
-import com.flash3388.flashlib.vision.processing.analysis.Analyser;
-import com.flash3388.flashlib.vision.processing.analysis.Analysis;
+import com.flash3388.flashlib.vision.analysis.Analyser;
+import com.flash3388.flashlib.vision.analysis.Analysis;
 
 import java.util.Optional;
 import java.util.function.Consumer;
